Variable Explorer: Move display_to_value from dicteditor to utils
